Module alloy_serde::num

Helpers for dealing with numbers. Numeric serde helpers.

Modules

u64_hex

serde functions for handling u64 as [U64]

u64_hex_opt

serde functions for handling Option<u64> as [U64]

u64_hex_or_decimal

serde functions for handling primitive u64 as [U64]

u64_hex_or_decimal_opt

serde functions for handling primitive optional u64 as [U64]

u128_hex_or_decimal

serde functions for handling primitive u64 as [U64]

Structs

U64HexOrNumber

A u64 wrapper type that deserializes from hex or a u64 and serializes as hex.

Enums

NumberOrHexU256

An enum that represents either a serde_json::Number integer, or a hex [U256].

Functions

from_int_or_hex

Deserializes the input into a U256, accepting both 0x-prefixed hex and decimal strings with arbitrary precision, defined by serde_json’s Number.

from_int_or_hex_opt

Deserializes the input into an Option<U256>, using from_int_or_hex to deserialize the inner value.