Crate alloy_serde

alloy-serde

Serde related helpers for Alloy.

Re-exports

pub use storage::JsonStorageKey;

pub use ttd::*;

Modules

displayfromstr

Serde functions for (de)serializing using FromStr and Display

quantity

Serde functions for encoding primitive numbers using the Ethereum JSON-RPC “quantity” format.

storage

Storage related helpers.

ttd

Serde functions for encoding the TTD using a Geth compatible format.

Structs

OtherFields

Generic type for capturing additional fields when deserializing structs.

WithOtherFields

An extension to a struct that allows to capture additional fields when deserializing.

Functions

null_as_default

For use with serde’s deserialize_with on a sequence that must be deserialized as a single but optional (i.e. possibly null) value.

reject_if_some

For use with serde’s deserialize_with on a field that must be missing.

serialize_b256_hex_string_no_prefix

Serialize a B256 as a hex string without the “0x” prefix.

serialize_hex_string_no_prefix

Serialize a byte vec as a hex string without the “0x” prefix.